Add eslint-config and update Pulumi settings for user-service
This commit involves adding 'eslint-config' in package.json and a corresponding lock file. Simultaneously, it removes 'user-service.ts' from resources/Pulumi due to changes in the structure of files required for deploying the user-service. The modifications in package.json assist in keeping the JavaScript code clean and consistent, while the update in the Pulumi setup readjusts the service deployment pipeline.