Improve image handling and add image resizing
Added support for image retrieval by filename and added an image resizing utility. Enhanced the image model to include URL and enhanced the handling of images in the MongoDB GridFS. Also, added the ability to retrieve and resize the image based on width and height parameters passed in the request.