Add API and ethang-toolbelt topics to documentation
Two new detailed topics, API and ethang-toolbelt, were added to the documentation. These new topics give insight into the API setup and provide an overview of the @ethang/toolbelt library. Furthermore, the home.tree was updated to include references to these new topics, which enhance the comprehensibility of the documentation.