Update documentation with API and ethang-toolbelt topics
The documentation has been enhanced with the inclusion of two new detailed topics - API and ethang-toolbelt. These additions offer insights into the configuration of the API and introduce the @ethang/toolbelt library. References to these topics have also been added to the home.tree file to improve the overall readability and understandability of the documentation.